requirements.txt
fastapi
pydantic
uvicorn
pymongo
dnspython
python-dotenv
motor==2.5.1
isort
